requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Cython==0.29.14
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.2
et-xmlfile==1.0.1
gensim==3.8.3
idna==2.10
jdcal==1.4.1
jieba==0.42.1
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
nltk==3.5
numpy==1.20.1
openpyxl==3.0.6
pandas==1.2.3
Pillow==8.1.2
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
selenium==3.141.0
six==1.15.0
sklearn==0.0
smart-open==4.2.0
soupsieve==2.2
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.58.0
urllib3==1.26.3
whitenoise==5.2.0
wikipedia==1.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
psycopg2==2.8.6
